// Canteen/src/app/models/manage-catalog.model.ts
export interface CategoryDto {
    categoryId: number;
    category: string;
    description: string;
  }

export interface PositionDto {
    positionId: number;
    position: string;
}

export interface MembershipDto {
    memberShipId: number;
    membership: string;
    loyaltyPoints: string;
    status: string;
  }